2 commits at one time: vfs: define struct filename and have getname() return it
getname() is intended to copy pathname strings from userspace into a
kernel buffer. The result is just a string in kernel space. It would
however be quite helpful to be able to attach some ancillary info to
the string.

For instance, we could attach some audit-related info to reduce the
amount of audit-related processing needed. When auditing is enabled,
we could also call getname() on the string more than once and not
need to recopy it from userspace.

This patchset converts the getname()/putname() interfaces to return
a struct instead of a string. For now, the struct just tracks the
string in kernel space and the original userland pointer for it.

Later, we'll add other information to the struct as it becomes
convenient.

Signed-off-by: Jeff Layton <jlayton@redhat.com>
Signed-off-by: Al Viro <viro@zeniv.linux.org.uk>

AND

 vfs: embed struct filename inside of names_cache allocation if possible

In the common case where a name is much smaller than PATH_MAX, an extra
allocation for struct filename is unnecessary. Before allocating a
separate one, try to embed the struct filename inside the buffer first. If
it turns out that that's not long enough, then fall back to allocating a
separate struct filename and redoing the copy.

Signed-off-by: Jeff Layton <jlayton@redhat.com>
Signed-off-by: Al Viro <viro@zeniv.linux.org.uk>

Conflicts:
	arch/alpha/kernel/osf_sys.c
	arch/arm64/kernel/sys.c
	arch/arm64/kernel/sys_compat.c
	arch/avr32/kernel/process.c
	arch/blackfin/kernel/process.c
	arch/cris/arch-v10/kernel/process.c
	arch/cris/arch-v32/kernel/process.c
	arch/h8300/kernel/process.c
	arch/hexagon/kernel/syscall.c
	arch/ia64/kernel/process.c
	arch/m32r/kernel/process.c
	arch/microblaze/kernel/sys_microblaze.c
	arch/mips/kernel/linux32.c
	arch/mips/kernel/syscall.c
	arch/openrisc/kernel/process.c
	arch/parisc/hpux/fs.c
	arch/parisc/kernel/process.c
	arch/parisc/kernel/sys_parisc32.c
	arch/score/kernel/sys_score.c
	arch/sh/kernel/process_32.c
	arch/sh/kernel/process_64.c
	arch/sparc/kernel/process_32.c
	arch/sparc/kernel/process_64.c
	arch/sparc/kernel/sys_sparc32.c
	arch/tile/kernel/process.c
	arch/unicore32/kernel/sys.c
	fs/namei.c
	include/linux/fs.h