qup_i2c: Add runtime PM support to I2C QUP controller
Runtime PM is used to turn clocks on and off when the
controller is not in use. When the controller is used
for messaging, pm_runtime_get_sync is called before
sending the message. pm_runtime_mark_last_busy and
pm_runtime_put_autosuspend are called when message
is sent and/or response is received.

Change-Id: I970def8eb601f7beb8ac256ed1496ddd48ca1257
CRs-Fixed: 406910
Signed-off-by: Alok Chauhan <alokc@codeaurora.org>