qseecom: Allow a generic client to vote for clks
There are some generic clients that will need to enable and
disable clks.  We need to account for these clients accessing
the CE clks and disable them according when the handle is
released.

This is done by making the perf_enabled, fast_load_enabled flags
part of the GENERIC handle.

Change-Id: I572f1a41768f500371bacf39ac959c51e0095efc
Signed-off-by: Mona Hossain <mhossain@codeaurora.org>