qseecom: Propagate error to process incomplete cmd.
When an error to process an INCOMPLETE command is encountered,
qseecom driver returns the error to HLOS client thereby leaving
the secure app that requested service from listener hanging.

This results in further communication to QSEE to be blocked.
In order to unblock the QSEE, so that it can accept new commands
from HLOS client, we need to return the status of error back to
QSEE to propagate it to secure app.

Change-Id: I69251bcc39f44759c0fd9c4b64e48a824c3ec80b
Signed-off-by: Mona Hossain <mhossain@codeaurora.org>